pyproject.toml
[tool.poetry]
name = "caracara"
version = "0.1.0"
description = ""
authors = ["Vadim Zifra"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^4.2.13"
django-environ = "^0.11.2"
djangorestframework = "^3.13.1"
django-two-factor-auth = {extras = ["phonenumberslite"], version = "^1.15.0"}
django-rest-durin = "^1.1.0"
django-extensions = "^3.2.1"
django-cryptography = "^1.1"
celery = "^5.4.0"
docker = "^7.0.0"
redis = "^4.5.1"
gunicorn = "^22.0.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
pre-commit = "^2.21.0"
pytest-factoryboy = "^2.5.1"
freezegun = "^1.2.2"
django-webtest = "^1.9.10"
ruff = "^0.4.8"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "B0",   # flake8-bugbear
    "E",    # pycodestyle Error
    "F",    # Pyflakes
    "I",    # isort
    "DJ",   # flake8-django
]
ignore = [
    "E501", # line-too-long
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
